Change the non-portable unsigned long to uintptr_t
The unsigned long size in gcc on 64-bit Linux is 64-bit while on MinGW
on 64-bit Windows is 32-bit (compiled with MinGW). This not only cause
compilation error but also failing tests on MinGW.

+++ m4ri/io.c
@@ -299,8 +299,8 @@ mzd_t *mzd_from_jcf(const char *fn, int verbose) {
   FILE *fh   = fopen(fn, "r");
 
   rci_t m, n;
-  long p       = 0;
-  long nonzero = 0;
+  intptr_t p       = 0;
+  intptr_t nonzero = 0;
 
   if (!fh) {
     if (verbose) printf("Could not open file '%s' for reading\n", fn);
@@ -326,8 +326,8 @@ mzd_t *mzd_from_jcf(const char *fn, int verbose) {
 
   A = mzd_init(m, n);
 
-  long i = -1;
-  long j = 0;
+  intptr_t i = -1;
+  intptr_t j = 0;
 
   while (fscanf(fh, "%ld\n", &j) == 1) {
     if (j < 0) { i++, j = -j; }
